Breaking out page task
